+++ src/ContaoManager/Plugin.php
@@ -45,10 +45,6 @@ class Plugin implements BundlePluginInterface, ConfigPluginInterface
 
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {
-        if (class_exists('HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle')) {
-            $loader->load('@HeimrichHannotFlatpickrBundle/Resources/config/config_encore.yml');
-        }
-
 		$loader->load('@HeimrichHannotFlatpickrBundle/Resources/config/services.yml');
 	}
 }